+++ tests/tools/test_calc_overlap_fc.py
@@ -1,17 +1,19 @@
 """Test functions related to files containing genomic ranges."""
 
-import sctoolbox.tools as tools
-import sctoolbox.utils as utils
+import sctoolbox.tools as tl
+import sctoolbox.utils as ul
 import os
 import pytest
 import anndata as ad
 import pkg_resources
 
+# ---------------------------- FIXTURES -------------------------------- #
+
 
 @pytest.fixture
 def bam():
     """Load bam file."""
-    bam_dir = os.path.join(os.path.dirname(__file__), 'data', 'atac')
+    bam_dir = os.path.join(os.path.dirname(__file__), '..', 'data', 'atac')
     bam_path = os.path.join(bam_dir, 'mm10_atac.bam')
 
     return bam_path
@@ -31,7 +33,7 @@ def gtf():
 @pytest.fixture
 def gtf_with_header():
     """Load a gtf like file with a header."""
-    gtf_path = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'gtf_testdata', 'cropped_gencode.v41.unsorted.gtf')
+    gtf_path = os.path.join(os.path.dirname(__file__), '..', 'data', 'atac', 'gtf_testdata', 'cropped_gencode.v41.unsorted.gtf')
 
     return gtf_path
 
@@ -39,7 +41,7 @@ def gtf_with_header():
 @pytest.fixture
 def fragments():
     """Load a bed file containing fragments."""
-    fragments_path = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'mm10_sorted_fragments.bed')
+    fragments_path = os.path.join(os.path.dirname(__file__), '..', 'data', 'atac', 'mm10_sorted_fragments.bed')
 
     return fragments_path
 
@@ -47,7 +49,7 @@ def fragments():
 @pytest.fixture
 def bed():
     """Load a bed with blacklisted regions."""
-    bed_path = os.path.join(os.path.dirname(__file__), 'data', 'atac', 'hg38.blacklist.v2_sorted.bed')
+    bed_path = os.path.join(os.path.dirname(__file__), '..', 'data', 'atac', 'hg38.blacklist.v2_sorted.bed')
 
     return bed_path
 
@@ -55,10 +57,13 @@ def bed():
 @pytest.fixture
 def adata():
     """Load adata."""
-    adata_dir = os.path.join(os.path.dirname(__file__), 'data', 'atac')
+    adata_dir = os.path.join(os.path.dirname(__file__), '..', 'data', 'atac')
     adata = ad.read_h5ad(adata_dir + '/mm10_atac.h5ad')
 
     return adata
+
+
+# --------------------------- Tests --------------------------------- #
 
 
 @pytest.mark.parametrize("out", [None, 'tmp'])
@@ -67,7 +72,7 @@ def test_convert_gtf_to_bed(tmpdir, gtf, out):
     if out:
         out = str(tmpdir)
 
-    sorted_bed, temp = tools._convert_gtf_to_bed(gtf, out=out)
+    sorted_bed, temp = tl._convert_gtf_to_bed(gtf, out=out)
     name = gtf + "_sorted.bed"
 
     if out:
@@ -77,7 +82,7 @@ def test_convert_gtf_to_bed(tmpdir, gtf, out):
 
     assert sorted_bed == expected and os.path.isfile(sorted_bed)
 
-    utils.rm_tmp(temp_files=temp)
+    ul.rm_tmp(temp_files=temp)
 
 
 @pytest.mark.parametrize("regions_file", ['bed', 'gtf'])
@@ -94,11 +99,11 @@ def test_fc_fragments_in_regions(tmpdir, adata, bed, gtf, bam, fragments, region
     if fragments_file == 'fragments':
         fragments_file = fragments
 
-    tools.fc_fragments_in_regions(adata,
-                                  regions_file=regions_file,
-                                  bam_file=bam_file,
-                                  fragments_file=fragments_file,
-                                  regions_name='promoters',
-                                  temp_dir=str(tmpdir))
+    tl.fc_fragments_in_regions(adata,
+                               regions_file=regions_file,
+                               bam_file=bam_file,
+                               fragments_file=fragments_file,
+                               regions_name='promoters',
+                               temp_dir=str(tmpdir))
 
     assert 'fold_change_promoters_fragments' in adata.obs.columns